refac ui: organize component module
Colors, fonts and icons are components of the
design system, they are moved to the corresponding
module. For clarity, text and badge modules are moved
in their own files.
