+++ src/ui/component/icon.rs
@@ -2,7 +2,7 @@ use iced::{Font, HorizontalAlignment, Length, Text};
 
 const ICONS: Font = Font::External {
     name: "Icons",
-    bytes: include_bytes!("../../static/icons/bootstrap-icons.ttf"),
+    bytes: include_bytes!("../../../static/icons/bootstrap-icons.ttf"),
 };
 
 fn icon(unicode: char) -> Text {
