+++ src/ui/component/image.rs
@@ -1,6 +1,6 @@
 use iced::widget::svg::{Handle, Svg};
 
-const LOGO: &'static [u8] = include_bytes!("../../static/images/revault-colored-logo.svg");
+const LOGO: &'static [u8] = include_bytes!("../../../static/images/revault-colored-logo.svg");
 
 pub fn revault_colored_logo() -> Svg {
     let h = Handle::from_memory(LOGO.to_vec());
