+++ src/ui/mod.rs
@@ -1,10 +1,6 @@
 pub mod app;
-mod color;
 mod component;
 mod error;
-mod font;
-mod icon;
-pub mod image;
 mod menu;
 mod message;
 mod state;
